# routes/goals.js
const express = require('express');
const router = express.Router();
const { authenticateJWT } = require('../middleware/auth');

// Get all goals
router.get('/', authenticateJWT, async (req, res) => {
  try {
    // TODO: Implement goals retrieval
    const goals = [
      {
        id: '1',
        title: 'Complete Project Alpha',
        description: 'Finish the main features of Project Alpha',
        userId: req.user._id,
        targetDate: '2024-01-31',
        progress: 75,
        status: 'in_progress'
      },
      {
        id: '2',
        title: 'Learn React Native',
        description: 'Master React Native development',
        userId: req.user._id,
        targetDate: '2024-03-15',
        progress: 30,
        status: 'in_progress'
      }
    ];
    
    res.json({
      message: 'Goals retrieved successfully',
      goals
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch goals', error: err.message });
  }
});

// Get user goals
router.get('/user/:id', authenticateJWT, async (req, res) => {
  try {
    const { id } = req.params;
    
    // TODO: Implement user goals retrieval
    const userGoals = [
      {
        id: '1',
        title: 'Complete Project Alpha',
        description: 'Finish the main features of Project Alpha',
        userId: id,
        targetDate: '2024-01-31',
        progress: 75,
        status: 'in_progress'
      }
    ];
    
    res.json({
      message: 'User goals retrieved successfully',
      userId: id,
      goals: userGoals
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch user goals', error: err.message });
  }
});

// Create new goal
router.post('/', authenticateJWT, async (req, res) => {
  try {
    const { title, description, targetDate, priority, category } = req.body;
    
    // TODO: Implement goal creation
    const newGoal = {
      id: Date.now().toString(),
      title,
      description,
      userId: req.user._id,
      targetDate,
      priority: priority || 'medium',
      category: category || 'general',
      progress: 0,
      status: 'not_started',
      createdAt: new Date().toISOString()
    };
    
    res.status(201).json({
      message: 'Goal created successfully',
      goal: newGoal
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to create goal', error: err.message });
  }
});

// Update goal
router.put('/:id', authenticateJWT, async (req, res) => {
  try {
    const { id } = req.params;
    const updates = req.body;
    
    // TODO: Implement goal update
    const updatedGoal = {
      id,
      ...updates,
      updatedAt: new Date().toISOString()
    };
    
    res.json({
      message: 'Goal updated successfully',
      goal: updatedGoal
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to update goal', error: err.message });
  }
});

// Delete goal
router.delete('/:id', authenticateJWT, async (req, res) => {
  try {
    const { id } = req.params;
    
    // TODO: Implement goal deletion
    
    res.json({
      message: 'Goal deleted successfully',
      goalId: id
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to delete goal', error: err.message });
  }
});

// Update goal progress
router.patch('/:id/progress', authenticateJWT, async (req, res) => {
  try {
    const { id } = req.params;
    const { progress, notes } = req.body;
    
    // TODO: Implement goal progress update
    const progressUpdate = {
      goalId: id,
      progress,
      notes,
      updatedAt: new Date().toISOString(),
      status: progress >= 100 ? 'completed' : 'in_progress'
    };
    
    res.json({
      message: 'Goal progress updated successfully',
      progressUpdate
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to update goal progress', error: err.message });
  }
});

module.exports = router; 