# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Task Manager API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the AI Task Manager SaaS platform with task management, AI features, team management, gamification, analytics, and more."
  },
  "servers": [
    { "url": "http://localhost:3000/api" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/kpis": {
      "get": {
        "tags": ["KPIs"],
        "summary": "Get all KPIs (employer) or own KPIs (employee)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of KPIs." },
          "401": { "description": "Unauthorized." }
        }
      },
      "post": {
        "tags": ["KPIs"],
        "summary": "Create a new KPI (employer only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "targetValue": { "type": "number" },
                  "assignedTo": { "type": "string" },
                  "period": { "type": "string", "enum": ["weekly", "monthly", "quarterly", "yearly"] }
                },
                "required": ["name", "targetValue", "assignedTo"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "KPI created." },
          "400": { "description": "Validation error." },
          "401": { "description": "Unauthorized." }
        }
      }
    },
    "/kpis/{id}/progress": {
      "patch": {
        "tags": ["KPIs"],
        "summary": "Update KPI progress (employee only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentValue": { "type": "number" }
                },
                "required": ["currentValue"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "KPI progress updated." },
          "404": { "description": "KPI not found." },
          "401": { "description": "Unauthorized." }
        }
      }
    },
...
    "/kpis/{id}": {
      "get": {
        "tags": ["KPIs"],
        "summary": "Get KPI by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "KPI retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "KPI not found" },
          "500": { "description": "Internal server error" }
        }
      },
      "put": {
        "tags": ["KPIs"],
        "summary": "Update KPI (employer only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "targetValue": { "type": "number" },
                  "assignedTo": { "type": "string" },
                  "period": { "type": "string", "enum": ["weekly", "monthly", "quarterly", "yearly"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "KPI updated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "KPI not found" },
          "500": { "description": "Internal server error" }
        }
      },
      "delete": {
        "tags": ["KPIs"],
        "summary": "Delete KPI (employer only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "KPI deleted successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "KPI not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/kpis/calculate": {
      "post": {
        "tags": ["KPIs"],
        "summary": "Calculate KPI metrics",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kpiId": { "type": "string", "description": "KPI ID to calculate" },
                  "userId": { "type": "string", "description": "User ID for calculation" },
                  "timeframe": { "type": "string", "enum": ["current", "previous", "custom"], "description": "Timeframe for calculation" }
                },
                "required": ["kpiId", "userId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "KPI calculation completed successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "KPI not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/goals": {
      "get": {
        "tags": ["Goals"],
        "summary": "Get all goals",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Goals retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["Goals"],
        "summary": "Create new goal",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "description": "Goal title" },
                  "description": { "type": "string", "description": "Goal description" },
                  "targetDate": { "type": "string", "format": "date", "description": "Target completion date" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"], "description": "Goal priority" },
                  "category": { "type": "string", "description": "Goal category" }
                },
                "required": ["title", "description", "targetDate"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Goal created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/goals/user/{id}": {
      "get": {
        "tags": ["Goals"],
        "summary": "Get user goals",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User goals retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/goals/{id}": {
      "put": {
        "tags": ["Goals"],
        "summary": "Update goal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "targetDate": { "type": "string", "format": "date" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"] },
                  "category": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Goal updated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      },
      "delete": {
        "tags": ["Goals"],
        "summary": "Delete goal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Goal deleted successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/goals/{id}/progress": {
      "patch": {
        "tags": ["Goals"],
        "summary": "Update goal progress",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "progress": { "type": "number", "minimum": 0, "maximum": 100, "description": "Progress percentage" },
                  "notes": { "type": "string", "description": "Progress notes" }
                },
                "required": ["progress"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Goal progress updated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  }
}